Move iot Client mock
